rl_module/requirements.txt
attrs==22.1.0
click==8.1.3
coloredlogs==15.0.1
flasgger==0.9.5
Flask==2.2.2
flatbuffers==2.0.7
humanfriendly==10.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.15.0
MarkupSafe==2.1.1
mistune==2.0.4
mpmath==1.2.1
numpy==1.23.2
onnxruntime==1.12.1
packaging==21.3
protobuf==4.21.5
pyparsing==3.0.9
pyrsistent==0.18.1
PyYAML==6.0
six==1.16.0
sympy==1.11.1
torch==1.12.1
typing_extensions==4.3.0
Werkzeug==2.2.2
zipp==3.8.1
